Improved screen adaptability & Added devicelist activity
The screen adaptability adds potential for scrolling if there is no
chance everything will fit on the screen.

The devicelist activity will display all devices in the local area in
the form of a list with characteristics. Oce of these will then be
selected to enter the brew screen for.